5758 - Organize meta/assessment/validation/description folder
