+++ src/server/cache/repository/validation/description/updateDescriptionLinkValidations.ts
@@ -3,7 +3,7 @@ import { Assessment } from 'meta/assessment/assessment'
 import { Cycle } from 'meta/assessment/cycle'
 import { SectionName } from 'meta/assessment/section'
 import { RecordDescriptionValidations } from 'meta/assessment/validation/description'
-import { DescriptionValidations } from 'meta/assessment/validation/descriptionValidations'
+import { DescriptionValidations } from 'meta/assessment/validation/descriptionValidations/descriptionValidations'
 import { Objects } from 'utils/objects'
 
 import { getKeyCountry, Keys } from 'server/cache/repository/keys'
